Moved file utils classes to separate directory
