+++ src/main/java/com/txdk/files/FileReader.java
@@ -1,4 +1,4 @@
-package com.txdk.frontend;
+package com.txdk.files;
 
 import java.io.IOException;
 import java.io.InputStream;
