+++ src/main/java/com/txdk/files/IconResizer.java
@@ -1,4 +1,4 @@
-package com.txdk.frontend;
+package com.txdk.files;
 
 import java.awt.Image;
